control: make file names more consistent, add data points processor
