+++ bench-configuration/01-write-1M-keys-client-variable/02-zookeeper-r3.4.9.yaml
@@ -38,7 +38,7 @@ step2:
     connections: 1
     clients: 1
     # if specified, connections, clients are overwritten
-    connections_clients: [1, 3, 5, 10, 50, 100, 500, 700, 1000]
+    connections_clients: [1, 10, 50, 100, 300, 500, 700, 1000]
     key_size: 256
     value_size: 1024
     total_requests: 1000000
