+++ control/step2_stress_util_client_io.go
